Refactor get storage options function
