+++ AGENTS.md
@@ -81,7 +81,7 @@ After individual CAIRO runs complete, post-processing scripts consolidate result
 | **`rate_design/`**                  | Package root. Heat pump rate design lives under `rate_design/hp_rates/`.                                                                                                                                                                                                                                                                                                                                                                                            |
 | **`rate_design/hp_rates/`**         | Shared scenario entrypoint (`run_scenario.py`), shared **Justfile** (primary task interface for all states). State-specific thin Justfiles and `config/` dirs live under `rate_design/hp_rates/{ny,ri}/`.                                                                                                                                                                                                                                                           |
 | **`rate_design/hp_rates/{ny,ri}/`** | State-specific thin Justfile (imports shared), `state.env`, and `config/` (tariffs JSON in tariffs/electric and tariffs/gas, tariff_maps CSV in tariff_maps/electric and tariff_maps/gas, marginal_costs). Large artifacts (buildstock raw/processed, cairo_cases) are git-ignored; sync via S3 or keep local.                                                                                                                                                      |
-| **`data/eia/hourly_loads/`**        | EIA zone load fetch and utility load aggregation; eia_region_config (state/utility config, get_aws_storage_options); Justfile for fetch-zone-data and aggregate-utility-loads.                                                                                                                                                                                                                                                                                      |
+| **`data/eia/hourly_loads/`**        | EIA zone load fetch and utility load aggregation; eia_region_config (state/utility config); Justfile for fetch-zone-data and aggregate-utility-loads.                                                                                                                                                                                                                                                                                                               |
 | **`data/eia/861/`**                 | EIA-861 utility stats (PUDL yearly sales); fetch_electric_utility_stat_parquets.py; Justfile build-utility-stats (local parquet), update (upload to s3://data.sb/eia/861/electric_utility_stats/), fetch-utility-stats STATE (CSV to stdout).                                                                                                                                                                                                                       |
 | **`data/fred/cpi/`**                | FRED CPI series; Justfile fetch-cpi (local parquet/), upload (sync to s3://data.sb/fred/cpi/).                                                                                                                                                                                                                                                                                                                                                                      |
 | **`data/aspe/fpl/`**                | ASPE Federal Poverty Guidelines fetch; Justfile fetch. Output: utils/post/data/fpl_guidelines.yaml (used by LMI discount logic).                                                                                                                                                                                                                                                                                                                                    |
