+++ utils/mid/compute_seasonal_discount_inputs.py
@@ -9,7 +9,7 @@ from pathlib import Path
 from cloudpathlib import S3Path
 from dotenv import load_dotenv
 
-from data.eia.hourly_loads.eia_region_config import get_aws_storage_options
+from utils.file_io import get_aws_storage_options
 from utils.mid.compute_subclass_rr import (
     DEFAULT_BAT_METRIC,
     DEFAULT_GROUP_COL,
