+++ utils/post/apply_ny_lmi_discounts_to_cairo_bills.py
@@ -29,7 +29,7 @@ import polars as pl
 from cloudpathlib import S3Path
 from dotenv import load_dotenv
 
-from data.eia.hourly_loads.eia_region_config import get_aws_storage_options
+from utils.file_io import get_aws_storage_options
 from utils.post.lmi_common import (
     assign_ny_tier_expr,
     fpl_pct_expr,
