+++ utils/post/apply_ny_lmi_to_master_bills.py
@@ -41,7 +41,7 @@ from typing import Any, cast
 import polars as pl
 from dotenv import load_dotenv
 
-from data.eia.hourly_loads.eia_region_config import get_aws_storage_options
+from utils.file_io import get_aws_storage_options
 from utils.post.lmi_common import (
     assign_ny_tier_expr,
     fpl_pct_expr,
