+++ utils/pre/rev_requirement/analyze_rr_scatter.py
@@ -25,7 +25,7 @@ from plotnine import (
     theme_minimal,
 )
 
-from data.eia.hourly_loads.eia_region_config import get_aws_storage_options
+from utils.file_io import get_aws_storage_options
 from utils.scenario_config import get_residential_customer_count_from_utility_stats
 
 UtilityData = dict[str, str | float]
