+++ utils/pre/rev_requirement/compute_rr.py
@@ -389,7 +389,7 @@ def main() -> None:
     storage_options = None
     if args.path_electric_utility_stats.startswith("s3://"):
         try:
-            from data.eia.hourly_loads.eia_region_config import get_aws_storage_options
+            from utils.file_io import get_aws_storage_options
 
             storage_options = get_aws_storage_options()
         except ImportError:
